Optimized matterbridgePlatform
