+++ src/matterbridgeAccessoryPlatform.ts
@@ -23,7 +23,6 @@
 
 import { Matterbridge, PlatformConfig } from './matterbridge.js';
 import { MatterbridgePlatform } from './matterbridgePlatform.js';
-import { MatterbridgeDevice } from './matterbridgeDevice.js';
 import { AnsiLogger } from 'node-ansi-logger';
 
 /**
@@ -43,59 +42,5 @@ export class MatterbridgeAccessoryPlatform extends MatterbridgePlatform {
     this.type = 'AccessoryPlatform';
 
     this.log.debug(`Matterbridge${this.type} loaded`);
-  }
-
-  /**
-   * This method must be overridden in the extended class.
-   * It is called when the platform is started.
-   * Use this method to create the MatterbridgeDevice and call this.registerDevice().
-   * @param {string} [reason] - The reason for starting.
-   * @throws {Error} - Throws an error if the method is not overridden.
-   */
-  async onStart(reason?: string) {
-    this.log.error('Plugins must override onStart.', reason);
-    throw new Error('Plugins must override onStart.');
-  }
-
-  /**
-   * This method can be overridden in the extended class.
-   * It is called after the platform has been commissioned.
-   * Use this method to perform any configuration of your devices.
-   */
-  async onConfigure() {
-    this.log.debug("**The plugin doesn't override onConfigure.");
-  }
-
-  /**
-   * This method can be overridden in the extended class.
-   * It is called when the platform is shutting down.
-   * Use this method to clean up any resources.
-   * @param {string} [reason] - The reason for shutting down.
-   */
-  async onShutdown(reason?: string) {
-    this.log.debug("The plugin doesn't override onShutdown.", reason);
-  }
-
-  /**
-   * Registers a device with the Matterbridge platform.
-   * @param {MatterbridgeDevice} device - The device to register.
-   */
-  async registerDevice(device: MatterbridgeDevice) {
-    await this.matterbridge.addDevice(this.name, device);
-  }
-
-  /**
-   * Unregisters a device registered with the Matterbridge platform.
-   * @param {MatterbridgeDevice} device - The device to unregister.
-   */
-  async unregisterDevice(device: MatterbridgeDevice) {
-    await this.matterbridge.removeBridgedDevice(this.name, device);
-  }
-
-  /**
-   * Unregisters all devices registered with the Matterbridge platform.
-   */
-  async unregisterAllDevices() {
-    await this.matterbridge.removeAllBridgedDevices(this.name);
   }
 }
